+++ anybox/buildbot/openerp/__init__.py
@@ -3,8 +3,11 @@ import os
 import warnings
 from configurator import BuildoutsConfigurator
 
-def configure_from_buildouts(buildmaster_path, config):
+def configure_from_buildouts(buildmaster_path, config, **kw):
     """Load the configuration with what's needed for the buildouts.
+
+    explicit manifest_paths (iterable of paths to be interpreted from the
+    buildmaster can be passed)
     """
 
     if os.path.isfile(buildmaster_path):
@@ -14,5 +17,5 @@ def configure_from_buildouts(buildmaster_path, config):
             "is deprecated and will disappear in the future. Please "
             "pass the buildmaster dir instead (``basedir`` in ``master.cfg`` "
             "local variables).", DeprecationWarning)
-        buildmasterpath = os.path.split(buildmaster_path)[0]
-    BuildoutsConfigurator(buildmaster_path).populate(config)
+        buildmaster_path = os.path.split(buildmaster_path)[0]
+    BuildoutsConfigurator(buildmaster_path, **kw).populate(config)