+++ anybox/buildbot/openerp/buildouts.py
@@ -1,9 +1,7 @@
-import os
 from ConfigParser import ConfigParser
-from ConfigParser import NoOptionError
 
-def parse_manifest(buildmaster_dir):
+def parse_manifest(filepath):
     """Return a ConfigParser for buildouts MANIFEST, ready for queries."""
     parser = ConfigParser()
-    parser.read(os.path.join(buildmaster_dir, 'buildouts', 'MANIFEST.cfg'))
+    parser.read(filepath)
     return parser