+++ CHANGES.rst
@@ -1,11 +1,19 @@
 Changes
 ~~~~~~~
 
+0.8 (Unreleased)
+----------------
+ - 
+
 0.7
---
+---
  - launchpad #999069: Test run parts of build factories are now customizable.
+ - launchpad #1040070: can read several manifest files
+ - launchpad #1050842: now standalone buildouts paths are relative to manifest
+   directory.
+ - db_template buildout option.
  - launchpad #999066: Utility script to find a free port in a range
- - ignore divergences bzr branch pulls (notably for mirrors)
+ - ignore divergences in bzr branch pulls (notably for mirrors)
 
 0.6
 ---
@@ -18,6 +26,7 @@ Changes
    (buildout) by capability.
  - launchpad #1004916: slaves max_builds and notify_on_missing
     parameters now taken into account
+
 0.5
 ---
  - using vcs-clear-retry option of OpenERP recipe